refactoring utilities
